refactoring client
