+++ test-tools/edi-test-client/src/client/run_client.py
@@ -5,39 +5,9 @@ import logging
 from concurrent.futures import ThreadPoolExecutor
 sys.path.insert(0, str(Path(__file__).parent.parent))
 from config import load_settings
+import reqs
 import time
 
-
-
-SAMPLE_270 = (
-"ISA*00*          *00*          *ZZ*1234567        *ZZ*11111          *170508*1141*^*00501*000000101*1*P*:~\n"
-"GS*HC*XXXXXXX*XXXXX*20170617*1741*101*X*005010X279A1~\n"
-"ST*270*1234*005010X279A1~\n"
-"BHT*0022*13*10001234*20060501*1319~\n"
-"HL*1**20*1~\n"
-"NM1*PR*2*ABC COMPANY*****PI*842610001~\n"
-"HL*2*1*21*1~\n"
-"NM1*1P*2*BONE AND JOINT CLINIC*****SV*2000035~\n"
-"HL*3*2*22*0~\n"
-"TRN*1*93175-012547*9877281234~\n"
-"NM1*IL*1*SMITH*ROBERT****MI*11122333301~\n"
-"DMG*D8*19430519~\n"
-"DTP*291*D8*20060501~\n"
-"EQ*30~\n"
-"SE*13*1234~\n"
-"GE*1*101~\n"
-"IEA*1*000000101~\n"
-)
-
-
-def send_270_request(endpoint):
-    start = time.perf_counter()
-    try:
-        resp = requests.post(url=endpoint, data=SAMPLE_270)
-        elapsed = (time.perf_counter() - start)*1000
-        return resp.status_code, elapsed
-    except Exception:
-        return -1, (time.perf_counter() - start)*1000
 
 def log_result(fut, logger):
     status, elapsed = fut.result()
@@ -56,7 +26,7 @@ def main():
         next_run = time.perf_counter()
         try:
             while True:
-                fut = pool.submit(send_270_request, cfg.endpoint)
+                fut = pool.submit(reqs.send_270_request, cfg.endpoint)
                 fut.add_done_callback(lambda f: log_result(f, logger))
                 reqs_sent += 1
                 next_run += 1.0 / cfg.rps
